fix: ISS-765 default BitmapFont + ISS-770 Align at API boundaries (wave-D-L)
ISS-765: port the no-arg BitmapFont() and BitmapFont(flip) constructors
(BitmapFont.java:71-88), backed by the default 15pt Liberation Sans font
now shipped at sge/utils/lsans-15.{fnt,png} in sge resources. Both ctors
delegate to the 4-arg FileHandle ctor via Sge().files.classpath; placed
after their delegate target so the self-constructor invocation resolves.

ISS-770 (design pin): retype alignment params from raw Int to sge.utils.Align
at the API boundaries — GlyphLayout.setText/ctors/alignRuns,
BitmapFontCache.setText/addText, BitmapFont.draw — and drop the
Align<->Int .asInstanceOf/.toInt round-trips in the scene2d/ui bridges
(Label, SelectBox, SgeList, TextArea, TextField). Mechanical .toInt removals
in the g2d test suites that call the retyped boundary.

SgeTestFixture: doc note (also forces Zinc to re-evaluate the typeCheckErrors
red suites, whose macro-string dependency on BitmapFont is untracked).

+++ sge/src/main/scala/sge/scenes/scene2d/ui/SelectBox.scala
@@ -265,7 +265,7 @@ class SelectBox[T](initialStyle: SelectBox.SelectBoxStyle)(using Sge) extends Wi
 
   protected def drawItem(batch: Batch, font: BitmapFont, item: T, x: Float, y: Float, width: Float): GlyphLayout = {
     val string = itemToString(item)
-    font.draw(batch, string, x, y, 0, string.length(), width, alignment.toInt, false, Nullable("..."))
+    font.draw(batch, string, x, y, 0, string.length(), width, alignment, false, Nullable("..."))
   }
 
   /** Sets the alignment of the selected item in the select box. See {@link #getList()} and {@link List#setAlignment(int)} to set the alignment in the list shown when the select box is open.
